Prevent deletion of PVC that is referenced by an active pod
The important logic is inside pv_controller_base.go:257 inside
the deleteClaim method and I added some comments about the logic.

In order to discover if a PVC's PV is referenced by an active pod
I needed to add a podLister. This resulted in lots of small changes,
especially to the test infrastructure.